Add battle locations to raid cards.
