+++ src/routes/tables/raid-upgrade-material-card.tsx
@@ -3,6 +3,8 @@ import React from 'react';
 import { RarityMapper } from '@/fsd/5-shared/model/mappers/rarity.mapper';
 import { UnitShardIcon } from '@/fsd/5-shared/ui/icons';
 
+import { ICampaignBattleComposed } from '@/fsd/4-entities/campaign/@x/upgrade';
+import { CampaignLocation } from '@/fsd/4-entities/campaign/campaign-location';
 import { CharactersService } from '@/fsd/4-entities/character';
 import { UpgradeImage, UpgradesService } from '@/fsd/4-entities/upgrade';
 
@@ -12,6 +14,7 @@ interface Props {
     currentQuantity: number;
     desiredQuantity: number;
     relatedCharacterSnowprintIds: string[];
+    locations: ICampaignBattleComposed[];
 }
 
 export const RaidUpgradeMaterialCard: React.FC<Props> = ({
@@ -19,6 +22,7 @@ export const RaidUpgradeMaterialCard: React.FC<Props> = ({
     currentQuantity,
     desiredQuantity,
     relatedCharacterSnowprintIds,
+    locations,
 }) => {
     const rewardIcon = () => {
         const upgrade = UpgradesService.getUpgrade(upgradeMaterialSnowprintId);
@@ -46,21 +50,30 @@ export const RaidUpgradeMaterialCard: React.FC<Props> = ({
     };
 
     return (
-        <div className="flex-box item-start">
-            <div className="flex-box column">
-                {rewardIcon()}
-                <span>
-                    {currentQuantity}/{desiredQuantity}
-                </span>
-            </div>
-            <div className="flex flex-col gap-2">
-                <div className="flex flex-wrap gap-1">
-                    {relatedCharacterSnowprintIds.map(id => (
-                        <div key={id}>
-                            <UnitShardIcon icon={CharactersService.getUnit(id)?.roundIcon ?? id} height={30} />
-                        </div>
-                    ))}
+        <div className="w-full max-w-[400px] overflow-hidden p-[5px] [box-shadow:1px_2px_3px_rgba(0,_0,_0,_0.6)]">
+            <div className="flex-box item-start">
+                <div className="flex-box column">
+                    {rewardIcon()}
+                    <span>
+                        {currentQuantity}/{desiredQuantity}
+                    </span>
                 </div>
+                <div className="flex flex-col gap-2">
+                    <div className="flex flex-wrap gap-1">
+                        {relatedCharacterSnowprintIds.map(id => (
+                            <div key={id}>
+                                <UnitShardIcon icon={CharactersService.getUnit(id)?.roundIcon ?? id} height={30} />
+                            </div>
+                        ))}
+                    </div>
+                </div>
+            </div>
+            <div className="flex-box gap2 wrap">
+                {locations
+                    .filter(x => x.isSuggested && x.isUnlocked)
+                    .map(loc => {
+                        return <CampaignLocation key={loc.id} location={loc} short={true} unlocked={true} />;
+                    })}
             </div>
         </div>
     );
